EDU-3818 refactored code to move coaching es models to utilitites package

* EDU-3818 refactored code to move coaching es models to utilitites package

* EDU-3818 changed the version of utilities to 0.0.8

* fields were added part of other commit, lost in merging

* added no args constructor as per review comments

Approved-by: Shivam Singhal <shivam.singhal@paytmmall.com>
Approved-by: Divyanshu <divyanshu.sharma@paytmmall.com>

+++ digital-education-coaching/src/main/java/com/paytm/digital/education/coaching/consumer/model/response/search/CoachingCourseData.java
@@ -20,7 +20,7 @@ public class CoachingCourseData extends SearchBaseData {
     private String     logo;
     private String     courseType;
     private CourseType courseTypeEnum;
-    private Double     courseDurationDays;
+    private Integer     courseDurationDays;
     private Integer    duration;
     private String     durationType;
     private Integer    validity;
